Refactored architecture, readying app for tests
