+++ app/client.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import App from './components/App';
+
+React.render(
+  <App />,
+  document.getElementById('app')
+);
