+++ app/components/App.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Connect4 from './Connect4/Connect4';
+
+export default React.createClass({
+  render() {
+    return (
+      <div>
+        <h1>Connect Four</h1>
+        <Connect4 />
+      </div>
+    );
+  }
+});
