+++ app/components/Connect4/BoardStatus.js
