+++ package.json
@@ -1,31 +1,50 @@
 {
-  "name": "react-babel-sample",
-  "version": "0.0.1",
-  "description": "",
-  "main": "index.js",
+  "author": "Sergio Cruz",
+  "name": "react-connect4",
+  "version": "0.1.0",
+  "description": "Connect 4 game implemented in React",
+  "main": "app/client.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest",
+    "test-debug": "node-debug --nodejs --harmony ./node_modules/jest-cli/bin/jest.js --runInBand"
   },
-  "author": "",
   "license": "ISC",
   "dependencies": {
-    "babel": "^5.8.12",
-    "browserify": "^11.0.0",
     "classnames": "^2.1.3",
-    "react": "^0.13.3",
-    "reactify": "^1.1.1"
+    "react": "^0.13.3"
   },
   "devDependencies": {
+    "babel": "^5.8.12",
+    "babel-jest": "^5.3.0",
     "babelify": "^6.1.3",
     "browser-sync": "^2.8.0",
     "browserify": "^11.0.0",
-    "eslint": "^0.24.1",
+    "eslint": "^1.0.0",
     "eslint-plugin-react": "^3.1.0",
     "exorcist": "^0.4.0",
     "gulp": "^3.9.0",
     "gulp-sass": "^2.0.4",
     "gulp-util": "^3.0.6",
+    "jest-cli": "^0.4.17",
+    "reactify": "^1.1.1",
     "vinyl-source-stream": "^1.1.0",
     "watchify": "^3.3.0"
+  },
+  "jest": {
+    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
+    "testFileExtensions": [
+      "es6",
+      "js"
+    ],
+    "moduleFileExtensions": [
+      "js",
+      "json",
+      "es6"
+    ],
+    "unmockedModulePathPatterns": ["react"]
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/sergiocruz/react-connect4.git"
   }
 }
